rearranged some code, and wip version check
